+++ src/main/resources/beyond/momentours/mapper/like/LikeMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="beyond.momentours.like.query.mapper.LikeMapper">
+<mapper namespace="beyond.momentours.like.query.repository.LikeMapper">
 
     <select id="existsByMemberAndTarget" resultType="boolean">
         SELECT EXISTS (
